+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # changelog
 
+* add new `basicnew` package with this functions:
+  * `MapKeys`: MapKeys generate a new slice with all keys of a map.
+  * `MapValues`: MapValues generate a new slice with all values of a map.
+
 ## v0.6.0
 
 * :warning: **BREAKING CHANGES**: remove deprecated functions in basiccheck: `StringInSlice`, `EqualStringSlice`, `IntInSlice`, `Int64InSlice`, `OneOfStringsWith`, `AllStringsWith`